pyproject.toml
[project]
name = "victron-dlms-reader"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "gurux-common==1.0.13",
    "gurux-dlms==1.0.185",
    "gurux-net==1.0.23",
    "gurux-serial==1.0.21",
    "lxml>=5.0.0,<6.0.0",
    "pyyaml~=6.0.2",
]
